+++ src/lifetime_bot/__init__.py
@@ -6,6 +6,7 @@ An automated bot for reserving classes at Life Time.
 __version__ = "1.0.0"
 
 from lifetime_bot.api import LifetimeAPIClient
+from lifetime_bot.auth import DirectAPIAuthenticator
 from lifetime_bot.bot import LifetimeReservationBot
 from lifetime_bot.config import BotConfig, EmailConfig, SMSConfig
 from lifetime_bot.errors import LifetimeAPIError
@@ -15,9 +16,12 @@ from lifetime_bot.models import (
     RegistrationResult,
     SessionTokens,
 )
+from lifetime_bot.reservations import ReservationService
 
 __all__ = [
     "LifetimeReservationBot",
+    "DirectAPIAuthenticator",
+    "ReservationService",
     "BotConfig",
     "EmailConfig",
     "SMSConfig",
